requirements.txt
streamlit==1.42.2
google-genai==1.2.0